+++ composer.json
@@ -40,6 +40,9 @@
         "phpunit/phpunit": "^8.5"
     },
     "autoload": {
+        "files": [
+            "Helper/GlobalHelper.php"
+        ],
         "psr-4": {
             "Jonassiewertsen\\StatamicButik\\": "src"
         }
