+++ src/Http/Livewire/CartIcon.php
@@ -2,7 +2,7 @@
 
 namespace Jonassiewertsen\StatamicButik\Http\Livewire;
 
-use Jonassiewertsen\StatamicButik\Helpers\Cart as ShoppingCart;
+use Jonassiewertsen\StatamicButik\Helper\Cart as ShoppingCart;
 use Livewire\Component;
 
 class CartIcon extends Component
